+++ HISTORY.rst
@@ -10,6 +10,7 @@ picows Release History
 * Enable sse2, avx2, avx512 speedups for windows builds
 * Disconnect if unknown frame opcode is received (RFC 6455 requirement)
 * Disconnect if client doesn't mask frame or server mask (RFC 6455 requirement)
+* Do not log "Protocol writing pause/resume requested" if user overrides WSListener pause_writing/resume_writing
 
 1.18.0 (2026-04-03)
 ------------------